Added internal artifacts for OTel spec log bridge api.
