+++ src/OpenTelemetry.Api/Internal/Shims/IsExternalInit.cs
