+++ src/OpenTelemetry.Api/Logs/LoggerProviderBuilder.cs
@@ -1,0 +1,42 @@
+// <copyright file="LoggerProviderBuilder.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+#nullable enable
+
+namespace OpenTelemetry.Logs;
+
+/// <summary>
+/// LoggerProviderBuilder base class.
+/// </summary>
+internal abstract class LoggerProviderBuilder
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="LoggerProviderBuilder"/> class.
+    /// </summary>
+    protected LoggerProviderBuilder()
+    {
+    }
+
+    /// <summary>
+    /// Adds instrumentation to the provider.
+    /// </summary>
+    /// <typeparam name="TInstrumentation">Type of instrumentation class.</typeparam>
+    /// <param name="instrumentationFactory">Function that builds instrumentation.</param>
+    /// <returns>Returns <see cref="LoggerProviderBuilder"/> for chaining.</returns>
+    public abstract LoggerProviderBuilder AddInstrumentation<TInstrumentation>(
+        Func<TInstrumentation> instrumentationFactory)
+        where TInstrumentation : class;
+}
